+++ test/index.test.ts
@@ -1,14 +1,17 @@
 import { use } from 'chai';
 import * as cap from 'chai-as-promised';
 
+import { suite as ArrayValidatorTests } from './tests/arrayValidator.test';
 import { suite as BigUserTest } from './tests/biguser.test';
 import { suite as customNameTests } from './tests/customName.test';
 import { suite as IndexTests } from './tests/dbIndex.test';
 import { suite as DefaultClassesTests } from './tests/dClasses.test';
 import { suite as ErrorTests } from './tests/errors.test';
 import { suite as GCFDTest } from './tests/getClassForDocument.test';
+import { suite as GlobalTest } from './tests/globalOptions.test';
 import { suite as HookTest } from './tests/hooks.test';
 import { suite as Inheritance } from './tests/inheritance.test';
+import { suite as OverwrittenModels } from './tests/overwrittenModel.test';
 import { suite as RefTest } from './tests/ref.test';
 import { suite as ShouldAddTest } from './tests/shouldAdd.test';
 import { suite as ShouldRunTests } from './tests/shouldRun.test';
@@ -17,11 +20,21 @@ import { suite as TypeguardsTest } from './tests/typeguards.test';
 
 import { connect, disconnect } from './utils/mongooseConnect';
 
+/*
+ * // use this style
+ * import { suite as ShouldAddTest } from './shouldAdd.test'
+ * ...
+ * describe('Should add', ShouldAddTest.bind(this));
+ * ...
+ */
+
 use(cap);
 
 describe('Typegoose', () => {
   before(connect);
   after(disconnect);
+
+  describe('Global Options', GlobalTest.bind(this));
 
   describe('BigUser', BigUserTest.bind(this));
 
@@ -35,9 +48,11 @@ describe('Typegoose', () => {
 
   describe('String Validators', StringValidatorTests.bind(this));
 
+  describe('Array Validators', ArrayValidatorTests.bind(this));
+
   describe('getClassForDocument()', GCFDTest.bind(this));
 
-  describe('Test if the correct errors are returned', ErrorTests.bind(this));
+  describe('Should Error', ErrorTests.bind(this));
 
   describe('Default Classes', DefaultClassesTests.bind(this));
 
@@ -48,4 +63,6 @@ describe('Typegoose', () => {
   describe('inheritance', Inheritance.bind(this));
 
   describe('customName', customNameTests.bind(this));
+
+  describe('Overwritten Model', OverwrittenModels.bind(this));
 });